+++ frontend/src/components/Bullet.module.css
@@ -1,0 +1,5 @@
+.selector {
+    /* Styles for the Bullet component */
+    margin-bottom: 20px;
+  }
+
