+++ frontend/src/components/Checkbox.module.css
@@ -1,0 +1,5 @@
+.checkboxGroup {
+    /* Styles for the Checkbox component */
+    margin-bottom: 20px;
+  }
+
