--- frontend/src/components/ExamsCheckboxGroup.module.css
@@ -1,5 +1,0 @@
-.checkboxGroup {
-    /* Styles for the ExamsCheckboxGroup component */
-    margin-bottom: 20px;
-  }
-
