--- frontend/src/components/ExperienceSelector.module.css
@@ -1,5 +1,0 @@
-.selector {
-    /* Styles for the ExperienceSelector component */
-    margin-bottom: 20px;
-  }
-
